requirements.txt
# Productivity Agent - Minimal Essential Requirements
# Core functionality with most compatible versions

# Core Python dependencies
python-dotenv>=1.0.0
requests>=2.25.0
aiohttp>=3.8.0
schedule>=1.1.0
typing-extensions>=4.0.0

# Data processing core
numpy>=1.21.0,<2.0.0
pandas>=1.3.0,<3.0.0

# Pydantic
pydantic>=2.0.0,<3.0.0

# LlamaIndex minimal
llama-index-core>=0.10.0
llama-index>=0.10.0

# Essential ML packages
transformers>=4.20.0
sentence-transformers>=2.2.0
torch>=1.9.0
huggingface-hub>=0.10.0

# Vector database
faiss-cpu>=1.7.0

# Composio minimal
composio-core>=0.3.0

# API integrations
PyGithub>=1.58.0
jira>=3.0.0
slack-sdk>=3.20.0

# Console output
rich>=13.0.0
colorama>=0.4.4
tabulate>=0.8.0

# Utilities
requests-oauthlib>=1.3.0
PyYAML>=6.0
click>=8.0.0
psutil>=5.8.0
